Fixed minor issue with benchmark common.h file.
D_MAX should be double and not float.

+++ mbench/common.h
@@ -4,4 +4,4 @@ enum {
 };
 
 const float F_MAX = 0.8123f;
-const float D_MAX = 0.8123;
+const double D_MAX = 0.8123;
